backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
anthropic==0.25.1
httpx==0.27.0
python-dotenv==1.0.1
aiohttp==3.9.5
beautifulsoup4==4.12.3
lxml==5.2.2
python-multipart==0.0.9
pytest==8.2.0
pytest-asyncio==0.23.7
